# docs/openapi.json
{
  "openapi": "3.0.0",
  "paths": {
...
    "/submissions": {
      "post": {
        "description": "Submit source code for judging. Returns the submission number for tracking via SSE.",
        "operationId": "SubmissionsController_create",
        "parameters": [],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/CreateSubmissionDto"
              }
            }
          }
        },
        "responses": {
          "201": {
            "description": "Submission accepted for judging"
          },
          "400": {
            "description": "Validation error"
          },
          "401": {
            "description": "Not authenticated"
          },
          "429": {
            "description": "Too many submissions"
          }
        },
        "security": [
          {
            "bearer": []
          }
        ],
        "summary": "Submit a solution",
        "tags": [
          "submissions"
        ]
      },
      "get": {
        "operationId": "SubmissionsController_list",
        "parameters": [
          {
            "name": "problemId",
            "required": false,
            "in": "query",
            "description": "Filter by problem ID",
            "schema": {
              "type": "string"
            }
          },
          {
            "name": "page",
            "required": false,
            "in": "query",
            "description": "Page number (default: 1)",
            "schema": {
              "type": "string"
            }
          },
          {
            "name": "limit",
            "required": false,
            "in": "query",
            "description": "Items per page (default: 20, max: 100)",
            "schema": {
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Paginated list of submissions"
          }
        },
        "security": [
          {
            "bearer": []
          }
        ],
        "summary": "List my submissions",
        "tags": [
          "submissions"
        ]
      }
    },
    "/submissions/{number}/stream": {
      "get": {
        "description": "Opens a Server-Sent Events stream that emits real-time updates as testcases are judged. The stream closes when judging completes.",
        "operationId": "SubmissionsController_stream",
        "parameters": [
          {
            "name": "number",
            "required": true,
            "in": "path",
            "description": "Submission number",
            "schema": {
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "SSE stream of judging events"
          },
          "401": {
            "description": "Not authenticated"
          },
          "404": {
            "description": "Submission not found"
          }
        },
        "security": [
          {
            "bearer": []
          }
        ],
        "summary": "Stream submission judging progress (SSE)",
        "tags": [
          "submissions"
        ]
      }
    },
    "/submissions/{number}": {
      "get": {
        "operationId": "SubmissionsController_detail",
        "parameters": [
          {
            "name": "number",
            "required": true,
            "in": "path",
            "description": "Submission number",
            "schema": {
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Submission detail with testcase results"
          },
          "401": {
            "description": "Not authenticated"
          },
          "404": {
            "description": "Submission not found"
          }
        },
        "security": [
          {
            "bearer": []
          }
        ],
        "summary": "Get submission detail",
        "tags": [
          "submissions"
        ]
      }
    },
...
  },
  "info": {
    "title": "EpochJudge API",
    "description": "纪元 — 开源 Online Judge 平台 API",
    "version": "0.1.0",
    "contact": {}
  },
...
  "servers": [],
  "components": {
    "securitySchemes": {
      "bearer": {
        "scheme": "bearer",
        "bearerFormat": "JWT",
        "type": "http"
      }
    },
    "schemas": {
...
      "CreateSubmissionDto": {
        "type": "object",
        "properties": {
          "problemId": {
            "type": "string"
          },
          "language": {
            "type": "string",
            "enum": [
              "JAVASCRIPT",
              "C",
              "CPP",
              "PYTHON",
              "JAVA"
            ]
          },
          "judgeMode": {
            "type": "string",
            "enum": [
              "OI",
              "ACM"
            ]
          },
          "sourceCode": {
            "type": "string"
          },
          "contestId": {
            "type": "string"
          }
        },
        "required": [
          "problemId",
          "language",
          "sourceCode"
        ]
      },
...
    }
  }
}